+++ src/app/components/Highlight/Highlight.examples.tsx
@@ -1,0 +1,6 @@
+import { Example } from '../examples';
+import Highlight from './Highlight';
+
+export const HighlightItem: Example = () => (
+  <Highlight events={['kill', 'assist']} timestamp={145} />
+);
