--- src/app/components/SummaryHighlightItem/SummaryHighlightItem.examples.tsx
@@ -1,6 +1,0 @@
-import { Example } from '../examples';
-import SummaryHighlightItem from './SummaryHighlightItem';
-
-export const SummaryHighlightsItem: Example = () => (
-  <SummaryHighlightItem events={['kill', 'assist']} timestamp={145} />
-);
